[FAB-2198] Gossip envelope refactoring
In the previous episode of [FAB-2198]:
	https://gerrit.hyperledger.org/r/#/c/5907/
	Adjust gossip membership layer
We adjusted the discovery layer and got rid of the usage of the protos
there.

Now, I'm gradually integrating the envelope with the gossip message and actually
making the signing and verification work on the raw payload instead of
on a computed payload that is non deterministic.

Also, the SignedEndpoint which was a part of the membership entity "Member"
is no more, and its functionality was moved outside of the GossipMessage,
to an external "Secret" message type.
This type will be used to hold parts of GossipMessage that the peers may want to
omit as they forward messages to peers that shouldn't get this information.

The current use-case for this, is FAB-2007 that enforces peers to not expose
the internal endpoints of peers in their own organization.
This data can't reside inside the GossipMessage anymore, because it is marshalled
into a payload and signed.
Therefore, we need to extract it into a side entity that will be
part of the Envelope that will be sent in gossip. T
hen, the peers can easily omit this envelope while
preserving the signature on the payload that the source peer produced.

In the (very-soon) future, I'll get rid of the coupling between the GossipMessage
and the Envelope reference inside of it.

Change-Id: Ib910cba1f69bd356174ceb64ee22e2a1d9d15cf5
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>