Removing security section from core.yaml
This change-set removes the security section from core.yaml
because not needed anymore.
The security configuration is demanded to the each module
that can decide how and where to take the configuration from.
An example, is the configuration of a channel that carries
information about the hash function to be used.

Change-Id: Ie2918bdb65413b35f207e40b1a6c8dac22f22a4d
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>