[FAB-2364] Create common orderer configupdate path
https://jira.hyperledger.org/browse/FAB-2364

The multichannel code has a special handling for config update messages,
but this same config update mechanism needs to be used for channel
reconfiguration.

This changeset adds a new set of common configuration update processing
code.  It does not hook it into the broadcast path, or remove the
duplicated functionality in the multichain package, so as to make this
CR easier to review.

Change-Id: Idc746c666a455bfba917ec88a1acbec0a2e1d33e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>