[FAB-2391] Create common config Proposer
https://jira.hyperledger.org/browse/FAB-2391

Today, each config handler implements the Propose/Commit/Rollback API on
its own, which has led to a lot of duplicated code.

This also means that each config handler must handle its own proto
unmarshaling and error handling which makes inspecting the config from a
human perspective cumbersome.  This is the first step in mapping the
config to/from a human readable form.

Change-Id: I1dfd595b1e8be2a63ef82cb0b91616a932d4e83b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>