[FAB-2396] Move orderer config to common Proposer
https://jira.hyperledger.org/browse/FAB-2396

This is the second in a series of CRs which will migrate the config
value handlers off of their own custom transaction handling code and
onto the common Proposer code.  This will not only reduce the amount of
reimplimented logic (and corresponding bugs) and will also facilitate
the two-way config translation for the configtxgen tool.

Change-Id: I0e4bd1228c12b02fbf074b05379e917d463bf877
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>