[FAB-2448] Add standard values proto initializer
https://jira.hyperledger.org/browse/FAB-2448

The proposer framework specifies a handler interface which requires
Allocate to create the backing resources necessary to handle
unmarshaling config and to map the config keys to the correct proto
structure.

This provides nice flexibility for arbitrary configuration, but, in
reality, all config is instantiated in almost the same way.

This CR adds a standard values proto intializer which will, given a
struct whose fields are proto messages, will build the config map based
on the field names, and allocate the memory necessary to unmarshal into
the struct.  This should help remove significant amounts of code as the
assorted config handlers are moved to the Proposer framework.

Change-Id: Ic54d116e8bbfa61dd6839068d6195026f215477f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>