Tune gossip default bootstrap and skip localhost conn
Currently, the peer/core.yaml is configured to connect to 0.0.0.0:7051
This makes gossip emit annoying warnings about unable to connect to such
an address to the logs.
I changed the default to be 127.0.0.1:7051

On top of that, there is no reason to connect to a bootstrap address
which is yourself. Also made the bootstrap connection process
skip such an endpoint.

Added a test that checks the filtering function I introduced,
and also added a test in the discovery layer that fails if a connection
to yourself is being attempted.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
Change-Id: Ifeaa04aaf385cd21b6671c94441d7e3612ea59e0