[FAB-2335] Move channel config to common Proposer
https://jira.hyperledger.org/browse/FAB-2335

This is the first in a series of CRs which will migrate the config value
handlers off of their own custom transcation handling code and onto the
common Proposer code.  This will not only reduce the amount of
reimplimented logic (and corresponding bugs) but also eventually
facilitate two way config translation for the configtxgen tool.

Change-Id: I81b07e3d0fb322fe9f4397b8e70273a9155f8e58
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>