[FAB-2545] Add tool to create various crypto configs
https://jira.hyperledger.org/browse/FAB-2545

In order to get a test (or even real) system running,
there is a lot of cryptographic material required:
- root certificates for CAs (e.g. fabric-ca)
- MSPs for organizations running peers
- Local MSPs for peers
- MSPs for ordererer organizations
- Local MSPs for ordering nodes

This CR adds a tool named "cryptogen" which
will create these artifacts for you.  It allows
you to specify the number of peer organizations,
the number of peers per organization and the
number of ordering nodes (shims).  It currently
only creates a single orderer organization.

To run, "./cryptogen" and it will display the
command line options

Change-Id: I15f135dc2893f7492566eb8ac5d02b2f4963ccd3
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>