[FAB-2605] Fix hard-coded start for deliver_stdout
https://jira.hyperledger.org/browse/FAB-2605

The SeekInfo message broadcasted by the deliver_stdout client had its
start field hard-coded to the `SeekOldest` message. This isn't really a
problem so far since we use the client to deliver from the oldest block
available anyway, but the code is wrong, and when we extend the client
(as we will do in a subsequent changeset), leaving this unfixed will
come back to haunt us.

Change-Id: Id2b06be1a632e42104bd696f6c67f23da27b5a03
Signed-off-by: Kostas Christidis <kostas@christidis.io>