[FAB-2615] Remove Ingress/EgressPolicyNames refs
https://jira.hyperledger.org/browse/FAB-2615

Now that the hierarchical config and policies are in place, the old
style of specifying ingress and egress policies is wrong.

This CR switches the orderer to use the ChannelReaders and
ChannelWriters policies instead of the Ingress/Egress policy names and
removes the references to those now removed config policies.

Change-Id: Ie818bfc37bb5d2b4eb55addaf76ac510a2fdfd4b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>