[FAB-2007] Gossip/Comm deep probing
When a peer receives a join-channel event, it doesn't know whether
the anchor peers in the genesis block indeed match their said organizations.
A malicious attacker may create a channel and add an anchor peer of
its own in each organization, and this way- make peers expose
their internal endpoints which is not wanted.

This commit introduces another method - Handkshake() that
performs a handshake to the remote peer, and returns its identity,
or an error if either a communication or authentication error occures.
The certificate then can be used to extract the remote peers' organization.

Change-Id: I0a0e71a10cb5831f07c72be28308cadb486c1d3a
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>