[FAB-2658] Bug fix: pass correct block data to MCS
The MessageCryptoService complains:
verifyBlock -> WARN 0cc Received fabricated block from [...]\
 in DataUpdate: Failed casting SignedBlock to []byte on channel [foo]

Because the object that was passed wasn't a []byte as expected but
the object that contains the []byte as a field.
I changed the signature of VerifyBlock to accept []byte instead
of the previous interface type api.SignedBlock and removed the latter type.

Change-Id: I7320c398b117072c0790f77d9c0f9305b1adf5ea
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>