[FAB-2493] Package up GOLANG source precisely
This patch modifies the current GOLANG platform driver
for packaging the source code (used by the CLI, etc) to
be much more precise.  The current code more or less
grabs the entire GOPATH on the platform in which the
CLI is invoked.  The new code will only grab the bare
minimum of source files to build a specified package.

Fixes FAB-2493 for the fabric-peer component.

Change-Id: I048e4f654f2d13fe658dbe8153898020e2b2f4c7
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>