[FAB-2695] (PA) mv msp_principal.proto protos/msp
This is intended for (PA) post-alpha.

https://jira.hyperledger.org/browse/FAB-2695

Before the proto compilation was fixed to handle importing protos across
directories, the msp_principal.proto was put into protos/common/.  It
does not make sense here, as it is only MSP related, and it belongs in
protos/msp/.

Note, that is an API change only, no ABI is modified, so this should not
impact existing clients until the regenerate the protobufs.

Change-Id: I72f52354d76e25237ad4831e8e25d6c9772196f6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>