Add a "make test-cmd" target
This is a convenience function to allow a user to generate a
"go test" command string with the requisite build-specific
parameters already filled in.  It does not run any code.  It
is only for allowing the user to forumlate their own go-test
session (typically in conjunction with running a specific
pkg/test or under a debugger).

Change-Id: I3bfe9bfcaf136159b4a447a76a641b6ab0c54ce6
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>