[FAB-2714] Enable peer to start with TLS enabled
https://jira.hyperledger.org/browse/FAB-2714

Prior to this change, the peer did not honor the TLS
setting in core.yaml.  With this change, the peer will
now read the peer.tls.* settings to determine whether or
not to start the main peer server with TLS enabled.

This following changes were made:
- core/comm/testdata/certs/generate.go - needed to
add usage extension to be able to use server certs as
client certs

- core/peer/peer.go - added package scoped peerServer variable in
order to support dynamic updates from config transactions

- peer/node/start.go - loads TLS config and calls function in
core/peer package in order to create the peer server

The tests do include a test for mutual (client) TLS
authentication, but this is not yet enabled for the peer
on start up

Change-Id: I9a6d19b928c36b7500d004a57b7d76e40bb18fd9
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>