[FAB-2723] (PA) Remove old ChainCreationPolicyName
Before FAB-1302, channel creation was governed by a list of
ChainCreationPolicyNames.  This mechanism was replaced by the notion of
consortiums and a corresponding consortium channel creation policy.
This allows channel creation policies to be tied to the consortium,
rather than to an arbitrary list and is both simpler, and more
expressive.

Consequently, the old chain creaiton policy name, and chain creation
policy protos are no longer needed.  This CR removes them to erase any
confusion.  This is an ABI break, and will prevent channels created with
the older style config to fail to start.

Change-Id: I08f82d66a8bf48c2b3301ed8f1394d7e4edb53ad
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Signed-off-by: Kostas Christidis <kostas@christidis.io>