[FAB-2824]  Incorrect configtx.yaml selected
When using configtxgen locally, the configtx.yaml from the source
folder will always be selected even if Orderer or Peer config
paths are specified.

For loop in localconfig/congif.go is not exited when a valid config
path is found. This is inconsistent with how it is handled in
provisional/provisional.go.

Change-Id: I3d00b713fedc059e4a10de0e0197d4db3d3826c2
Signed-off-by: Nick Murray <nick@passkit.com>