[FAB-2936] Vendor latest version of sarama library
https://jira.hyperledger.org/browse/FAB-2936

This is the first in a series of changesets that will allow us to
quickly query the state of the Kafka cluster. This is itself needed so
that we can decide whether there's a point in repeating a certain
request (e.g. for the creation of a channel/partition in Kafka) or not.

We're vendoring the latest version of the sarama library because it
introduces the `Brokers()` method to the `Client` interface, and this
allows us to easily get the list of active brokers.

This changeset also adds a relevant (and now req'd) config setting, see:
https://github.com/Shopify/sarama/issues/816 for more.

Change-Id: Ia6bda8f237e087066f404f09f9d6ac359ed44b85
Signed-off-by: Kostas Christidis <kostas@christidis.io>