[FAB-2916] Refactor flogging package
There are several things on this package that could use improvement. Off
the top of my head as I'm writing this, I am noting down the following:

1. Test coverage is great but the tests are incredibly verbose and
filled with redundant checks (as an example: there is no need to ensure
that the log level is the default in
`TestSetModuleLevel_moduleWithSubmodule`, this is taken care of in a
previous test).
2. Asymmetry on the return types; some functions return the logging
level as the string type, others as the `logging.Level` type.
3. Functions whose purpose is unclear (`InitFromViper`), whose name is
inaccurate (`SetLoggingFormat` does not just set the logging format), or
whose logic is simply flawed (`InitFromViper` - pay attention to how the
input argument is used).
4. Outdated and --in some cases-- nonsensical comments on the exported
functions.

This is my attempt to tidy things up. I've also updated the code that
used to reference the now-deprecated methods.

Change-Id: I0cc020f20b3e4a4c2a4785f98aee2cb3282758bc
Signed-off-by: Kostas Christidis <kostas@christidis.io>