Enhance validation of chaincode names/versions in LCCC
This CR ensures chaincode names and versions only consist
of the allowable characters. Chaincode names should only consist of
alphanumerics, '_', and '-' while versions can also include '.' . It
also fixes the GetInstalledChaincodes query logic to account for the
fact that '.' is allowed in versions. Finally, this CR reworks some
of the test case logic for install, deploy, and upgrade to allow
for easier addition of standard test cases for those functions.

https://jira.hyperledger.org/browse/FAB-2922

Change-Id: I4ac29ed619eb91d2a04f4896546443ae81da432c
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>