[FAB-2756] Batch gossip state transfer request
Currently newly joined peer while syncing with rest of the network able
to send request to retrive entire ledger in one message, which is
in-efficient and presents potential exposure for denial of service
attacks.

This commits refactor gossip state trafer message in a way it
requires only to send two numbers representing the interval of required
missing blocks and makes sure this interval could not be longer than
specified by algorithm value.

Change-Id: I12cbf1922c39032125e1d168cea57eafa1013a32
Signed-off-by: Artem Barger <bartem@il.ibm.com>