[FAB-2980] Protobuf for ledger QueryResult
https://gerrit.hyperledger.org/r/#/c/7221/ added QueryResult
structures to chaincode shim.

This CR changes golang QueryResult structs such as KV (holds a result
of a range/execute query) and KeyModification (holds a result of a
history query) to protobuf messages. This makes the QueryResult
agnostic of language. As a result, both golang and java chaincode
shim can easily cast the QueryResult to appropriate struct depending
on the query type.

Further, TxID used in KeyModification struct is renamed to TxId to
to maintain consistency with ChannelHeader as both are exposed outside
fabric via Chaincode/Smart-Contract and Block, respectively.

Change-Id: I68a58d10a073ec8f83ed2523fc5ba0b82c61242e
Signed-off-by: senthil <cendhu@gmail.com>
Signed-off-by: Luis Sanchez <sanchezl@us.ibm.com>