[FAB-2997] don't invoke some scc thru proposals
This is the first of three fixes required for FAB-2997 (described in the
discussion for that jira item). In this change set we prevent
security-sensitive system chaincodes (such as escc or vscc) from being
invoked from the outside through proposals sent to the peer. A flag has
been added to the system chaincode struct to keep track of whether a system
chaincode can be invoked this way or not. A second flag was added to keep
track of whether the system chaincode can be invoked using a cc2cc
invocation. This flag will be used in the second change set related to this
jira item.

Change-Id: Iaf05217fcf7247f61fddb4a16afb0908c9735efc
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>