[FAB-2943] Event publishing failure fails block commit
The committer implementation reports an error if the publishing
of a block committing event failures due to an error that is related
to the eventhub server.

The method shouldn't return a failure, but only log a message to the log,
because if it returns failure- the state transfer module in gossip/state/state.go
doesn't update its metadata.

Change-Id: I134874654f1ce1453eb30939fedfa2c86c2634da
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>