[FAB-2928] link installation to instantiation [part-2]
This adds package specific signature to LCCC on instantiations.
It also prevents launching of chaincodes if the installed
chaincode does not match the signature of the instanitated chaincode.
In particular, it ensures future installs that don't comply with the
corresponding instantiation will be useless (cannot invoke them)..

Also refactored getChaincode function per Binh's suggestion
in Part-1.

Change-Id: I107c91a80002a336e42aff95d4aed7efe4141ead
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>