[FAB-2962] Gossip- Channel name obfuscation
This commit changes the stateInfo messages in gossip in the following way:
Instead of sending the plaintext of the channel in the channel field
of the GossipMessage - it now sends instead: Hash(channel || PKI-ID).

The added security benefit:
Given a message M with a channel C,
a peer not in C needs to do an exhaustive search / dictionary attack
on channel names, in order to guess C.
Without knowing C, it cannot ask an ordering service, or some other peer
to get a block for that channel because sending the hash would be worthless.

On reception of a StateInfo message from a peer with PKI-ID P, a peer
iterates over all channels (denote CHAN as the current iteration)
it participates in, and computes Hash(P || CHAN) and this way it
knows the channel the message relates to.
- If it is not in the channel, it would simply forward the message
  to peers in the organization (as done today).
- Else, it forwards the message to peers that their organization is a member
  of the channel (as done today).

Change-Id: I419ed3e440b936de80ec5624698b5d6e52376e98
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>