[FAB-3131] Peer hangs when CouchDB unresponsive
The peer hangs when the CouchDB network connection is unresponsive.
This can be corrected by adding a timeout to the network client
connection used for connecting to CouchDB.  If the connection
times out, then the request to CouchDB will return an error and
the retry logic will take over.

A new entry is added for CouchDB in core.yaml named connectionTimeout

couchDBConfig:
   couchDBAddress: 127.0.0.1:5984
   username:
   password:
   # Number of retries for CouchDB errors
   maxRetries: 3
   # Number of retries for CouchDB errors during peer startup
   maxRetriesOnStartup: 10
   # CouchDB connection timeout (unit: duration, e.g. 60s)
   connectionTimeout: 60s

This is a duration property configurable by the peer.

Change-Id: I5028029e7f303144465c2bfeebf540c5d8e64d82
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>