[FAB-3198] gossip: Un-necessary WARN log at shutdown
The gossip comm layer exposes a subscription-like API and
returns a dedicated channel to the caller.

Since it is convenient to iterate over a channel until it's closed,
or to understand that a channel has been closed when nil is returned
from select - it was implemented that at the comm layer shutdown -
the subscriptions channels are closed.

This might have caused a "send to closed channel" error-
and so- the implementation has a defer statement with recover()
that recovers from that.

However- it also logs a WARN message un-necessarily because that's not
something that shouldn't have happened.

Change-Id: I50d03eb8482e115782b3e4090330f72720883028
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>