[FAB-2094] Documenting MSP Setup & Best Practices
This changeset documents default MSP configuration parameters, and
its setup process on the peer and orderer side.

It also touches how MSP configuration updates take place within a channel, but
this is to be expanded after configuration updates related code is merged.

This changeset also discusses best practices and assumptions associated to
the default MSP considered by the fabric.

Change-Id: Ibe043202567619c7ad91f31ff7f1056b768db5e6
Signed-off-by: Elli Androulaki <lli@zurich.ibm.com>