[FAB-3297] Generate PKCS8 compliant EC keys
When BCCSP serialize EC private and public keys
to PEM format, it currently includes "ECDSA" in
the header.  While this is not illegal, it is
incompatible with libraries which expect PKCS8
format.

This change only uses PKCS8 for EC keys as most
libraries and tools support PKCS1 for RSA.

Needed to modify the based asn1 marshaling for
EC private keys as the Go implementation
includes an optional element which is
incompatible with several libraries with the
most important being the jsrsasign package
used by fabric-node-sdk project

Change-Id: I802dcdb28b46bd353efe1ed3868feb78f92e92c9
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>